+++ kaji/scripts/beta_soak_gate.py
@@ -19,7 +19,7 @@ BUDGETS = json.loads((ROOT / "kaji" / "benchmarks" / "beta-budgets.json").read_t
 ]
 EXPECTED_ARTIFACTS = {
     "python": "kaji_sdk-0.2.0b1-py3-none-any.whl",
-    "typescript": "kaji-sdk-0.2.0-beta.1.tgz",
+    "typescript": "kaji-sdk-0.2.0-beta.2.tgz",
 }