+++ kaji/scripts/validate_release_evidence.py
@@ -32,7 +32,7 @@ SEMVER = re.compile(
 )
 PYTHON_WHEEL = "kaji_sdk-0.2.0b1-py3-none-any.whl"
 PYTHON_SDIST = "kaji_sdk-0.2.0b1.tar.gz"
-TYPESCRIPT_TARBALL = "kaji-sdk-0.2.0-beta.1.tgz"
+TYPESCRIPT_TARBALL = "kaji-sdk-0.2.0-beta.2.tgz"
 TYPESCRIPT_GITHUB_TOOLS = (
     "add_comment",
     "create_issue",