+++ kaji/tests/test_tthw_composer.py
@@ -25,7 +25,7 @@ TIMINGS_TEMPLATE = (
 ARTIFACTS = {
     "kaji_sdk-0.2.0b1-py3-none-any.whl": ("python", "0.2.0b1"),
     "kaji_sdk-0.2.0b1.tar.gz": ("python", "0.2.0b1"),
-    "kaji-sdk-0.2.0-beta.1.tgz": ("typescript", "0.2.0-beta.1"),
+    "kaji-sdk-0.2.0-beta.2.tgz": ("typescript", "0.2.0-beta.2"),
 }